Move all backend state related app messages to own hook.
