+++ frontend/src/App.tsx
@@ -2,26 +2,9 @@ import { Outlet, useNavigate, useOutlet, useSearchParams } from "react-router"
 import StartupProgress from "./components/StartupProgress"
 import { useEffect, useState } from "react"
 import { useAppMessage } from "./lib/hooks/appMessage"
-import { Project, StatusBarUpdate } from "./types"
-import i18next from "i18next"
-import Settings from "./types/settings"
-import _ from "lodash"
-import { useProjectStore } from "./stores/projectStore"
+import { StatusBarUpdate } from "./types"
 import { MainMenu } from "./components/MainMenu"
-import { useRecentProjects, useSettingsStore } from "./stores/settingsStore"
-import { useControllerDefinitionsStore } from "./stores/definitionStore"
-import {
-  AuthenticationStatus,
-  BoardDefinitions,
-  ConnectedControllers,
-  ExecutionState,
-  HubHopState,
-  JoystickDefinitions,
-  MidiControllerDefinitions,
-  OverlayState,
-  ProjectStatus,
-  RecentProjects,
-} from "./types/messages"
+import { OverlayState } from "./types/messages"
 import {
   useKeyAccelerators,
   GlobalKeyAccelerators,
@@ -30,46 +13,21 @@ import LoaderOverlay from "./components/tables/config-item-table/LoaderOverlay"
 import { Toaster } from "./components/ui/sonner"
 import { useTheme } from "@/lib/hooks/useTheme"
 import { ToastNotificationHandler } from "./components/notifications/ToastNotificationHandler"
-import { useHubHopStateActions } from "./stores/stateStore"
 
-import testProject from "@/../tests/data/project.testdata.json" with { type: "json" }
-import testJsDefinition from "@/../tests/data/joystick.definition.json" with { type: "json" }
-import testMidiDefinition from "@/../tests/data/midicontroller.definition.json" with { type: "json" }
-import testRecentProjects from "@/../tests/data/recentProjects.testdata.json" with { type: "json" }
-import testControllers from "@/../tests/data/connectedControllers.testdata.json" with { type: "json" }
-
-import {
-  MidiControllerDefinition,
-  JoystickDefinition,
-} from "@/types/definitions"
 import DebugInfo from "@/components/DebugInfo"
-import { useExecutionStateStore } from "@/stores/executionStateStore"
-import { ProjectInfo } from "@/types/project"
-import { useControllerStore } from "@/stores/controllerStore"
 import { useTranslation } from "react-i18next"
-import { useAuth } from "react-oidc-context"
+import { useBackendStateAppMessages } from "@/lib/hooks/useBackendStateAppMessages"
 
 function App() {
+  // Initialize global app message handlers
+  useBackendStateAppMessages() 
   const [queryParameters] = useSearchParams()
   const navigate = useNavigate()
-  const { project, setProject, setProjectStatus } = useProjectStore()
-  const { setRecentProjects } = useRecentProjects()
-  const { setSettings } = useSettingsStore()
-  const { setControllers } = useControllerStore()
-  const {
-    setBoardDefinitions,
-    setJoystickDefinitions,
-    setMidiControllerDefinitions,
-  } = useControllerDefinitionsStore()
 
-  const { setIsRunning, setIsTesting } = useExecutionStateStore()
-
-  const setHubHopState = useHubHopStateActions()
   useKeyAccelerators(GlobalKeyAccelerators, true)
   const outlet = useOutlet()
   const [overlayVisible, setOverlayVisible] = useState(false)
   const { theme } = useTheme()
-  const auth = useAuth()
 
   // State for startup progress from app messages
   const [appStartupProgress, setAppStartupProgress] = useState<StatusBarUpdate>(
@@ -78,6 +36,16 @@ function App() {
       Text: "Starting...",
     },
   )
+
+  useAppMessage("StatusBarUpdate", (message) => {
+    setAppStartupProgress(message.payload as StatusBarUpdate)
+  })
+
+  useAppMessage("OverlayState", (message) => {
+    const overlayState = message.payload as OverlayState
+    console.log("OverlayState message received", overlayState)
+    setOverlayVisible(overlayState.Visible)
+  })
 
   const queryProgressValue = Number.parseInt(
     queryParameters.get("progress")?.toString() ?? "0",
@@ -92,97 +60,6 @@ function App() {
         }
       : appStartupProgress
 
-  useAppMessage("StatusBarUpdate", (message) => {
-    setAppStartupProgress(message.payload as StatusBarUpdate)
-  })
-
-  useAppMessage("Project", (message) => {
-    const project = message.payload as Project
-    console.log("Project message received", project)
-    setProject(project)
-  })
-
-  useAppMessage("RecentProjects", (message) => {
-    const recentProjects = message.payload as RecentProjects
-    setRecentProjects(recentProjects.Projects)
-    console.log("List of Recent Projects received", recentProjects.Projects)
-  })
-
-  useAppMessage("Settings", (message) => {
-    const settings = message.payload as Settings
-    console.log("Settings message received", settings)
-    setSettings(settings)
-
-    const language = settings.Language.split("-")[0]
-    if (!_.isEmpty(language)) i18next.changeLanguage(settings.Language)
-    else i18next.changeLanguage()
-  })
-
-  useAppMessage("BoardDefinitions", (message) => {
-    const boardDefinitions = message.payload as BoardDefinitions
-    console.log(
-      "BoardDefinitions message received",
-      boardDefinitions.Definitions,
-    )
-    setBoardDefinitions(boardDefinitions.Definitions)
-  })
-
-  useAppMessage("JoystickDefinitions", (message) => {
-    const joystickDefinitions = message.payload as JoystickDefinitions
-    console.log(
-      "JoystickDefinitions message received",
-      joystickDefinitions.Definitions,
-    )
-    setJoystickDefinitions(joystickDefinitions.Definitions)
-  })
-
-  useAppMessage("MidiControllerDefinitions", (message) => {
-    const definitions = message.payload as MidiControllerDefinitions
-    console.log(
-      "MidiControllerDefinitions message received",
-      definitions.Definitions,
-    )
-    setMidiControllerDefinitions(definitions.Definitions)
-  })
-
-  useAppMessage("ProjectStatus", (message) => {
-    const projectStatus = message.payload as ProjectStatus
-    console.log("ProjectStatus message received", projectStatus)
-    setProjectStatus(projectStatus)
-  })
-
-  useAppMessage("OverlayState", (message) => {
-    const overlayState = message.payload as OverlayState
-    console.log("OverlayState message received", overlayState)
-    setOverlayVisible(overlayState.Visible)
-  })
-
-  useAppMessage("HubHopState", (message) => {
-    const state = message.payload as HubHopState
-    setHubHopState(state)
-  })
-
-  useAppMessage("ConnectedControllers", (message) => {
-    const controllers = (message.payload as ConnectedControllers).Controllers
-    setControllers(controllers)
-  })
-
-  // Listen for auth state changes from C#
-  useAppMessage("AuthenticationStatus", async (message) => {
-    const authStatus = message.payload as AuthenticationStatus
-    console.log("AuthenticationStatus message received", authStatus)
-    try {
-      // Trigger silent signin to sync auth state from localStorage
-      if (authStatus.Authenticated) {
-        await auth.signinSilent()
-      } else {
-        auth.removeUser() // Clear user to ensure state is updated based on signinSilent result
-      }
-    } catch (err) {
-      console.log("Auth sync after window close:", err)
-    }
-  })
-
   useEffect(() => {
     if (startupProgress.Value == 100 && location.pathname == "/index.html") {
       console.log("Finished loading, navigating to home")
@@ -190,40 +67,6 @@ function App() {
     }
   }, [startupProgress.Value, navigate])
 
-  // this is only for easier UI testing
-  // while developing the UI
-  useEffect(() => {
-    if (
-      process.env.NODE_ENV === "development" &&
-      queryParameters.get("testdata") === "true" &&
-      !project // Only if no project loaded yet
-    ) {
-      setProject(testProject as Project)
-      setRecentProjects(testRecentProjects as ProjectInfo[])
-      setJoystickDefinitions([testJsDefinition as JoystickDefinition])
-
-      setMidiControllerDefinitions([
-        testMidiDefinition as MidiControllerDefinition,
-      ])
-
-      setControllers(testControllers as ConnectedControllers["Controllers"])
-    }
-  }, [
-    project,
-    queryParameters,
-    setJoystickDefinitions,
-    setMidiControllerDefinitions,
-    setProject,
-    setRecentProjects,
-    setControllers,
-  ])
-
-  useAppMessage("ExecutionState", (message) => {
-    console.log("ExecutionState message received", message.payload)
-    const { IsRunning, IsTesting } = message.payload as ExecutionState
-    setIsRunning(IsRunning)
-    setIsTesting(IsTesting)
-  })
   const { t } = useTranslation()
 
   return (
